refactor(db): one IsUniqueViolation for the Postgres callers
internal/api/admin.go and internal/user/user.go each spelled out the same
pgconn 23505 check. It now lives in internal/db, which both already import.
The SQLite variant in internal/frontdesk stays where it is: it inspects a
different driver error entirely. isForeignKeyViolation has no twin, so it
stays in internal/api.
