+++ internal/db/pgerr.go
@@ -1,0 +1,18 @@
+package db
+
+import (
+	"errors"
+
+	"github.com/jackc/pgx/v5/pgconn"
+)
+
+// IsUniqueViolation reports whether err is a PostgreSQL unique-constraint
+// violation (SQLSTATE 23505). Callers use it to turn a racing insert into a
+// 409/duplicate response instead of a 500. The SQLite Front Desk store has its
+// own variant: the driver reports a different error shape entirely.
+func IsUniqueViolation(err error) bool {
+	if pgErr, ok := errors.AsType[*pgconn.PgError](err); ok {
+		return pgErr.Code == "23505"
+	}
+	return false
+}
